notifications: cleanup, moving and refactoring
